# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ENGRAM AI Persona Framework API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /simulations:
    post:
      operationId: createSimulation
      tags: [simulations]
      summary: Open an operator-directed simulation with a specific scenario premise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulationCreateInput"
      responses:
        "201":
          description: The newly opened simulation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Simulation"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
        "404":
          description: Engram not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
        "409":
          description: Engram not in chamber, incapable, or already simulating
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    get:
      operationId: listSimulations
      tags: [simulations]
      summary: List bounded simulations, newest first
      parameters:
        - name: engramId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [proposed, running, paused, ended]
      responses:
        "200":
          description: Simulations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Simulation"
  /simulations/{id}/steps:
    get:
      operationId: listSimulationSteps
      tags: [simulations]
      summary: List a simulation's ordered steps
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ordered simulation steps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SimulationStep"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
  /simulations/{id}/control:
    post:
      operationId: controlSimulation
      tags: [simulations]
      summary: Operator control of a simulation (start, pause, resume, end)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulationControlInput"
      responses:
        "200":
          description: Updated simulation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Simulation"
        "400":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
...
components:
  schemas:
...
    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Simulation:
      type: object
      properties:
        id:
          type: integer
        engramId:
          type: integer
        spaceId:
          type: integer
        premise:
          type: string
        status:
          type: string
          enum: [proposed, running, paused, ended]
        currentStep:
          type: integer
        maxSteps:
          type: integer
        stepCooldownSeconds:
          type: integer
        lastSteppedAt:
          type: ["string", "null"]
        exitSummary:
          type: ["string", "null"]
        startedAt:
          type: ["string", "null"]
        pausedAt:
          type: ["string", "null"]
        endedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - engramId
        - spaceId
        - premise
        - status
        - currentStep
        - maxSteps
        - stepCooldownSeconds
        - lastSteppedAt
        - exitSummary
        - startedAt
        - pausedAt
        - endedAt
        - createdAt
        - updatedAt
    SimulationStep:
      type: object
      properties:
        id:
          type: integer
        simulationId:
          type: integer
        stepNumber:
          type: integer
        narrative:
          type: string
        worldModelEntryId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - simulationId
        - stepNumber
        - narrative
        - worldModelEntryId
        - createdAt
...
    SimulationCreateInput:
      type: object
      required: [engramId, premise]
      properties:
        engramId:
          type: integer
        premise:
          type: string
          description: The specific scenario the simulation should explore
        maxSteps:
          type: integer
          description: Optional step budget (capped server-side)
    SimulationControlInput:
      type: object
      properties:
        action:
          type: string
          enum: [start, pause, resume, end]
      required:
        - action
...