# lib/api-zod/src/generated/types/simulationCreateInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ENGRAM AI Persona Framework API
 * OpenAPI spec version: 0.1.0
 */

export interface SimulationCreateInput {
  engramId: number;
  /** The specific scenario the simulation should explore */
  premise: string;
  /** Optional step budget (capped server-side) */
  maxSteps?: number;
}